added scrolling, and window settings.
window state and geometry are saved to ini on exit, so now when you
resize and move the window it will be restored to those settings on next
launch.
Also scrolling support was added to help with smaller screen sizes and
for the ability to shrink the window so it doesn't take up as much room
on your desktop now.